apps/levantis-api/requirements.txt
Django>=5.0
psycopg2-binary
djangorestframework
django-cors-headers
drf-spectacular
redis
